requirements.txt
# Web Framework
flask
flask-restful
flask-cors

# Database
pymongo
redis

# Data Processing and ML
numpy
pandas
scikit-learn
scipy

# Geolocation
maxminddb
geoip2

# Utilities
python-dotenv
requests
user-agents
ua-parser

# Security
bcrypt
pyjwt

# Testing
pytest
